fix Wall...

+++ ft_memccpy.c
@@ -6,17 +6,17 @@
 /*   By: kshanti <kshanti@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/30 18:19:25 by kshanti           #+#    #+#             */
-/*   Updated: 2020/10/30 19:19:14 by kshanti          ###   ########.fr       */
+/*   Updated: 2020/10/30 19:28:15 by kshanti          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void		*ft_memccpy(void *s1, const void *s2, int c, size_t n)
+void			*ft_memccpy(void *s1, const void *s2, int c, size_t n)
 {
-	int		i;
-	char	*str1;
-	char	*str2;
+	size_t		i;
+	char		*str1;
+	char		*str2;
 
 	i = 0;
 	str1 = (char*)s1;
