+++ ft_strncmp.c
@@ -6,17 +6,17 @@
 /*   By: kshanti <kshanti@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/29 19:53:25 by kshanti           #+#    #+#             */
-/*   Updated: 2020/10/30 18:34:23 by kshanti          ###   ########.fr       */
+/*   Updated: 2020/10/30 19:30:49 by kshanti          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int			ft_strncmp(const char *s1, const char *s2, size_t n)
+int				ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int		i;
-	char	*str1;
-	char	*str2;
+	size_t		i;
+	char		*str1;
+	char		*str2;
 
 	i = 0;
 	str1 = (char*)s1;
